requirements.txt
fastapi
uvicorn
python-jose
streamlit
requests
aiohttp

